producer/requirements.txt
praw
kafka-python
requests
pytz
